custom: stop managing starship configuration files
The script that is the output of $(starship init zsh) has some quirks
that make it unsuitable to be ran from /etc/profile.d. It seems that the
scripts in this directory run before a few important are set (e.g.
ZSH_VERSION), which are expected by the starship init command.

Because of this, it was failing with:
/etc/profile.d/starship.sh:13: parse error: condition expected: $ZSH_VERSION

We are better off keeping the setup of starship in .zshrc and .bashrc.