custom: don't configure shell tools with system files
Although that approach worked for running shell tool on the host, it
also made it harder to run these tools inside toolbx (since the system
files inside the toolbx are different). It's better to keep these files
in the $HOME directory but we need figure out a good way to manage it.